Remove API calls and use hardcoded values
