+++ README.md
@@ -15,9 +15,6 @@ This app gets scores of users from an external API and displays it on the browse
 
 - Clone this repo to your machine by using Git clone: Git clone `https://github.com/GoddessNgozi/Leaderboard`
 - open in your vs code terminal and run live server
-- `npm install -g @lhci/cli@0.7.x`
-- `npm install --save-dev hint@6.x`
-- `npm install --save-dev stylelint@13.x stylelint-scss@3.x stylelint-config-standard@21.x stylelint-csstree-validator@1.x`
 
 ### Run the following tests:
